+++ libnetwork/iptables/firewalld.go
@@ -6,8 +6,11 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/pkg/rootless"
 	dbus "github.com/godbus/dbus/v5"
 	"github.com/pkg/errors"
 )
@@ -32,18 +35,35 @@ var (
 	connection *Conn
 
 	firewalldInitCalled bool
-	firewalldRunning    bool      // is Firewalld service running
-	onReloaded          []*func() // callbacks when Firewalld has been reloaded
+	firewalldRunning    bool // is Firewalld service running
+	// Time of the last firewalld reload, acquire firewalldReloadedAtMu before reading/writing.
+	firewalldReloadedAt   time.Time
+	firewalldReloadedAtMu sync.Mutex
+	// Mutex to serialise firewalld reload callbacks.
+	firewalldReloadMu sync.Mutex
+	onReloaded        []*func() // callbacks when Firewalld has been reloaded
 )
 
 // UsingFirewalld returns true if iptables rules will be applied via firewalld's
 // passthrough interface. The error return is non-nil if the status cannot be
 // determined because the initialisation function has not been called.
 func UsingFirewalld() (bool, error) {
-	if !firewalldInitCalled {
+	// If called before startup has completed, the firewall backend is unknown.
+	// But, if running rootless, the init function is not called because
+	// firewalld will be running in the host's netns, not in rootlesskit's.
+	if !firewalldInitCalled && !rootless.RunningWithRootlessKit() {
 		return false, fmt.Errorf("iptables.firewalld is not initialised")
 	}
 	return firewalldRunning, nil
+}
+
+// FirewalldReloadedAt returns the time at which the daemon last completed a
+// firewalld reload, or a nil-valued time.Time if it has not been reloaded
+// since the daemon started.
+func FirewalldReloadedAt() time.Time {
+	firewalldReloadedAtMu.Lock()
+	defer firewalldReloadedAtMu.Unlock()
+	return firewalldReloadedAt
 }
 
 // firewalldInit initializes firewalld management code.
@@ -145,9 +165,15 @@ func connectionLost() {
 
 // call all callbacks
 func reloaded() {
+	firewalldReloadMu.Lock()
+	defer firewalldReloadMu.Unlock()
 	for _, pf := range onReloaded {
 		(*pf)()
 	}
+
+	firewalldReloadedAtMu.Lock()
+	defer firewalldReloadedAtMu.Unlock()
+	firewalldReloadedAt = time.Now()
 }
 
 // OnReloaded add callback